refactor: bruk ny personkontrakt i VTP
Fjerner det gamle /rest/api/testscenarios/v2/opprett-endepunktet og tilhørende mapper. /personer blir eneste endepunkt for v2-personscenarioer.

Oppdaterer domene og mocks til String-baserte ISO alpha-3-landkoder, normaliserer alpha-2 ved innlesing og fjerner nv-i18n fra VTP-applikasjonen.
