+++ domene/src/main/java/no/nav/vtp/person/personopplysninger/GeografiskTilknytning.java
@@ -1,8 +1,6 @@
 package no.nav.vtp.person.personopplysninger;
 
-import com.neovisionaries.i18n.CountryCode;
-
-public record GeografiskTilknytning(CountryCode land, GeografiskTilknytningType type) {
+public record GeografiskTilknytning(String land, GeografiskTilknytningType type) {
 
     public enum GeografiskTilknytningType {
         BYDEL,
