+++ domene/src/main/java/no/nav/vtp/person/personopplysninger/Landkode.java
@@ -1,0 +1,9 @@
+package no.nav.vtp.person.personopplysninger;
+
+public final class Landkode {
+
+    public static final String NORGE = "NOR";
+
+    private Landkode() {
+    }
+}
