+++ domene/src/main/java/no/nav/vtp/person/personopplysninger/Medlemskap.java
@@ -2,11 +2,9 @@ package no.nav.vtp.person.personopplysninger;
 
 import java.time.LocalDate;
 
-import com.neovisionaries.i18n.CountryCode;
-
 public record Medlemskap(LocalDate fom,
                          LocalDate tom,
-                         CountryCode land,
+                         String land,
                          DekningsType trygdedekning) {
 
     public enum DekningsType {
