+++ domene/src/main/java/no/nav/vtp/person/personopplysninger/Statsborgerskap.java
@@ -1,6 +1,4 @@
 package no.nav.vtp.person.personopplysninger;
 
-import com.neovisionaries.i18n.CountryCode;
-
-public record Statsborgerskap(CountryCode land) {
+public record Statsborgerskap(String land) {
 }
