+++ server/src/main/java/no/nav/foreldrepenger/vtp/server/api/scenario/PersonMapper.java
@@ -1,39 +1,33 @@
 package no.nav.foreldrepenger.vtp.server.api.scenario;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
-import no.nav.foreldrepenger.vtp.kontrakter.person.AdresseDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.Adressebeskyttelse;
-import no.nav.foreldrepenger.vtp.kontrakter.person.ArbeidsavtaleDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.ArbeidsforholdDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.Arbeidsgiver;
-import no.nav.foreldrepenger.vtp.kontrakter.person.ArenaDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.ArenaSakerDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.ArenaVedtakDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.FamilierelasjonModellDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.GeografiskTilknytningDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.GrunnlagDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.InfotrygdDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.InntektYtelseType;
-import no.nav.foreldrepenger.vtp.kontrakter.person.InntektsperiodeDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.MedlemskapDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.OrganisasjonDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.PermisjonDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.Permisjonstype;
-import no.nav.foreldrepenger.vtp.kontrakter.person.PersonDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.PersonstatusDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.SigrunDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.SivilstandDto;
-import no.nav.foreldrepenger.vtp.kontrakter.person.Språk;
-import no.nav.foreldrepenger.vtp.kontrakter.person.StatsborgerskapDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.AdresseDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.ArbeidsavtaleDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.ArbeidsforholdDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.ArbeidsgiverDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.FamilierelasjonDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.GeografiskTilknytningDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.InntektsperiodeDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.Landkode;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.MedlemskapDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.OrganisasjonDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.PermisjonDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.PersonDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.PersonopplysningerDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.PersonstatusDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.PrivatArbeidsgiverDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.SkatteopplysningDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.SivilstandDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.StatsborgerskapDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.YtelseDto;
 import no.nav.vtp.person.Person;
 import no.nav.vtp.person.arbeidsforhold.Arbeidsavtale;
 import no.nav.vtp.person.arbeidsforhold.Arbeidsforhold;
@@ -57,39 +51,33 @@ import no.nav.vtp.person.personopplysninger.Rolle;
 import no.nav.vtp.person.personopplysninger.Sivilstand;
 import no.nav.vtp.person.personopplysninger.Statsborgerskap;
 import no.nav.vtp.person.skatt.Skatteopplysning;
-import no.nav.vtp.person.ytelse.LegacyKilde;
+
 import no.nav.vtp.person.ytelse.Ytelse;
 import no.nav.vtp.person.ytelse.YtelseType;
 
 public class PersonMapper {
 
+    private static final Organisasjon NAV_YTELSE_BETALING = new Organisasjon(
+            new Orgnummer("991013628"),
+            null,
+            new Organisasjon.Detaljer("NAV UTBETALING YTELSE", LocalDate.of(2003, 1, 1))
+    );
+
     private PersonMapper() {
-        /* This utility class should not be instantiated */
+        // utility
     }
 
     public static Person tilPerson(PersonDto p, Map<UUID, PersonIdent> identer, Optional<Person> eksisterende) {
-        var personopplysninger = tilPersonopplysninger(p, identer, eksisterende);
-        var arbeidsforhold = tilArbeidsforhold(p, identer);
-        var inntekt = tilInntekt(p, identer);
-        var ytelser = tilYtelser(p);
-        var skatteopplysninger = tilSkatteopplysnigner(p);
+        var personopplysninger = tilPersonopplysninger(p.personopplysninger(), identer, eksisterende);
+        var arbeidsforhold = tilArbeidsforhold(p.arbeidsforhold(), identer);
+        var inntekt = tilInntekt(p.inntekt(), p.ytelser(), identer);
+        var ytelser = tilYtelser(p.ytelser());
+        var skatteopplysninger = tilSkatteopplysninger(p.skatteopplysninger());
         return new Person(personopplysninger, arbeidsforhold, inntekt, ytelser, skatteopplysninger);
     }
 
-    private static List<Skatteopplysning> tilSkatteopplysnigner(PersonDto person) {
-        if (person.inntektytelse() == null || person.inntektytelse().sigrun() == null || person.inntektytelse().sigrun().inntektår() == null) {
-            return List.of();
-        }
-        return person.inntektytelse().sigrun().inntektår().stream()
-                .map(PersonMapper::tilSkatteopplysning)
-                .toList();
-    }
-
-    private static Skatteopplysning tilSkatteopplysning(SigrunDto.InntektsårDto inntektsår) {
-        return new Skatteopplysning(inntektsår.år(), inntektsår.beløp());
-    }
-
-    private static Personopplysninger tilPersonopplysninger(PersonDto p, Map<UUID, PersonIdent> identer, Optional<Person> eksisterende) {
+    private static Personopplysninger tilPersonopplysninger(PersonopplysningerDto p, Map<UUID, PersonIdent> identer,
+                                                            Optional<Person> eksisterende) {
         var navn = eksisterende
                 .map(e -> e.personopplysninger().navn())
                 .orElseGet(() -> generertTilfeldigNavn(p.kjønn()));
@@ -105,7 +93,7 @@ public class PersonMapper {
                 tilGeografiskTilknytning(p.geografiskTilknytning()),
                 tilFamilierelasjoner(p.familierelasjoner(), identer),
                 tilStatsborgerskap(p.statsborgerskap()),
-                tilSivilstand(p.sivilstand()),
+                tilSivilstander(p.sivilstand()),
                 tilPersonstatus(p.personstatus()),
                 tilMedlemskap(p.medlemskap()),
                 tilAdresser(p.adresser(), p.adressebeskyttelse()),
@@ -113,177 +101,21 @@ public class PersonMapper {
         );
     }
 
-    private static Rolle tilRolle(no.nav.foreldrepenger.vtp.kontrakter.person.Rolle rolle) {
-        return switch (rolle) {
-            case MOR -> Rolle.MOR;
-            case FAR -> Rolle.FAR;
-            case MEDMOR -> Rolle.MEDMOR;
-            case MEDFAR -> Rolle.MEDFAR;
-            case BARN -> Rolle.BARN;
-            case PRIVATE_ARBEIDSGIVER -> Rolle.PRIVATE_ARBEIDSGIVER;
-        };
+    private static Rolle tilRolle(no.nav.foreldrepenger.vtp.kontrakter.person.v2.Rolle rolle) {
+        if (rolle == null) {
+            return null;
+        }
+        return Rolle.valueOf(rolle.name());
     }
 
-    private static Navn generertTilfeldigNavn(no.nav.foreldrepenger.vtp.kontrakter.person.Kjønn kjønn) {
-        var generetNavn = no.nav.foreldrepenger.vtp.kontrakter.person.Kjønn.M.equals(kjønn) ? FiktivtNavn.getRandomMaleName() : FiktivtNavn.getRandomFemaleName();
+    private static Navn generertTilfeldigNavn(no.nav.foreldrepenger.vtp.kontrakter.person.v2.Kjønn kjønn) {
+        var generetNavn = no.nav.foreldrepenger.vtp.kontrakter.person.v2.Kjønn.M.equals(kjønn)
+                ? FiktivtNavn.getRandomMaleName()
+                : FiktivtNavn.getRandomFemaleName();
         return new Navn(generetNavn.fornavn(), null, generetNavn.etternavn());
     }
 
-    private static List<Arbeidsforhold> tilArbeidsforhold(PersonDto p, Map<UUID, PersonIdent> identer) {
-        if (p.inntektytelse() == null || p.inntektytelse().aareg() == null) {
-            return Collections.emptyList();
-        }
-
-        return Optional.ofNullable(p.inntektytelse().aareg().arbeidsforhold())
-                .orElse(Collections.emptyList())
-                .stream()
-                .map(dto -> tilArbeidsforhold(dto, identer))
-                .toList();
-    }
-
-    private static Arbeidsforhold tilArbeidsforhold(ArbeidsforholdDto dto, Map<UUID, PersonIdent> identer) {
-        return new Arbeidsforhold(
-                tilArbeidsgiver(dto.arbeidsgiver(), dto.arbeidsforholdId(), identer),
-                dto.ansettelsesperiodeFom(),
-                dto.ansettelsesperiodeTom(),
-                tilArbeidsforholdstype(dto.arbeidsforholdstype()),
-                tilArbeidsavtaler(dto.arbeidsavtaler()),
-                tilPermisjoner(dto.permisjoner())
-        );
-    }
-
-    private static no.nav.vtp.person.arbeidsforhold.Arbeidsgiver tilArbeidsgiver(Arbeidsgiver arbeidsgiver, String arbeidsforholdId,
-                                                                                 Map<UUID, PersonIdent> identer) {
-        if (arbeidsgiver instanceof OrganisasjonDto(no.nav.foreldrepenger.vtp.kontrakter.person.Orgnummer orgnummer, OrganisasjonDto.OrganisasjonsdetaljerDto organisasjonsdetaljer)) {
-            var detaljer = new Organisasjon.Detaljer(organisasjonsdetaljer.navn(), organisasjonsdetaljer.registreringsdato());
-            return new Organisasjon(new Orgnummer(orgnummer.value()), arbeidsforholdId, detaljer);
-        }
-        var uuid = ((no.nav.foreldrepenger.vtp.kontrakter.person.PrivatArbeidsgiver) arbeidsgiver).uuid();
-        return new PrivatArbeidsgiver(identer.get(uuid));
-    }
-
-    private static List<Inntektsperiode> tilInntekt(PersonDto p, Map<UUID, PersonIdent> identer) {
-        if (p.inntektytelse() == null || p.inntektytelse().inntektskomponent() == null) {
-            return Collections.emptyList();
-        }
-
-        return Optional.ofNullable(p.inntektytelse().inntektskomponent().inntektsperioder())
-                .orElse(Collections.emptyList())
-                .stream()
-                .map(dto -> tilInntektsperiode(dto, identer))
-                .toList();
-    }
-
-    private static Inntektsperiode tilInntektsperiode(InntektsperiodeDto dto, Map<UUID, PersonIdent> identer) {
-        return new Inntektsperiode(
-                tilArbeidsgiver(dto.arbeidsgiver(), null, identer),
-                dto.fom(),
-                dto.tom(),
-                dto.beløp(),
-                tilYtelseType(dto.inntektYtelseType()),
-                tilInntektFordel(dto.inntektFordel())
-        );
-    }
-
-    private static List<Ytelse> tilYtelser(PersonDto p) {
-        if (p.inntektytelse() == null) {
-            return Collections.emptyList();
-        }
-
-        var ytelser = new ArrayList<Ytelse>();
-
-        // Map Arena ytelser
-        if (p.inntektytelse().arena() != null && p.inntektytelse().arena().saker() != null) {
-            ytelser.addAll(tilYtelserFraArena(p.inntektytelse().arena()));
-        }
-
-        // Map Infotrygd ytelser
-        if (p.inntektytelse().infotrygd() != null && p.inntektytelse().infotrygd().ytelser() != null) {
-            ytelser.addAll(tilYtelserFraInfotrygd(p.inntektytelse().infotrygd()));
-        }
-
-        // Map Pesys ytelser
-        if (p.inntektytelse().pesys() != null && Boolean.TRUE.equals(p.inntektytelse().pesys().harUføretrygd())) {
-            ytelser.add(tilYtelserFraPesys());
-        }
-
-        return ytelser;
-    }
-
-    private static List<Ytelse> tilYtelserFraArena(ArenaDto arena) {
-        var arenaYtelser = new ArrayList<Ytelse>();
-        arena.saker().forEach(sak -> arenaYtelser.addAll(tilYtelseFraArenaVedtak(sak.tema(), sak.vedtak().getFirst())));
-        return arenaYtelser;
-    }
-
-    private static List<Ytelse> tilYtelseFraArenaVedtak(ArenaSakerDto.YtelseTema tema, ArenaVedtakDto vedtak) {
-        var ytelseType = tilYtelseTypeFraArenaTema(tema);
-
-        if (vedtak.meldekort() != null && !vedtak.meldekort().isEmpty()) {
-            // Bruk meldekort fom/tom direkte som Ytelse fom/tom
-            return vedtak.meldekort().stream()
-                    .map(mk -> new Ytelse(ytelseType, mk.fom(), mk.tom(), mk.dagsats(), mk.beløp(), null, LegacyKilde.ARENA, Collections.emptyList()))
-                    .toList();
-        }
-
-        return List.of(new Ytelse(ytelseType, vedtak.fom(), vedtak.tom(), vedtak.dagsats(), 0, null, LegacyKilde.ARENA, Collections.emptyList()));
-    }
-
-    private static YtelseType tilYtelseTypeFraArenaTema(ArenaSakerDto.YtelseTema tema) {
-        return switch (tema) {
-            case AAP -> YtelseType.ARBEIDSAVKLARINGSPENGER;
-            case DAG -> YtelseType.DAGPENGER;
-            default -> throw new IllegalStateException("Ikke støttet ytelse i arena!");
-        };
-    }
-
-    private static List<Ytelse> tilYtelserFraInfotrygd(InfotrygdDto infotrygd) {
-        return infotrygd.ytelser().stream()
-                .map(PersonMapper::tilYtelseFraInfotrygdGrunnlag)
-                .collect(Collectors.toList());
-    }
-
-    private static Ytelse tilYtelseFraInfotrygdGrunnlag(GrunnlagDto grunnlag) {
-        var ytelseType = tilYtelseTypeFraInfotrygdYtelse(grunnlag.ytelse());
-
-        // Calculate total utbetalingsgrad from vedtak
-        Integer totalUtbetalingsgrad = null;
-        if (grunnlag.vedtak() != null && !grunnlag.vedtak().isEmpty()) {
-            var avgUtbetalingsgrad = grunnlag.vedtak().stream()
-                    .map(GrunnlagDto.Vedtak::utbetalingsgrad)
-                    .filter(Objects::nonNull)
-                    .mapToInt(Integer::intValue)
-                    .average();
-            totalUtbetalingsgrad = avgUtbetalingsgrad.isPresent() ? (int) avgUtbetalingsgrad.getAsDouble() : null;
-        }
-
-        return new Ytelse(
-                ytelseType,
-                grunnlag.fom(),
-                grunnlag.tom(),
-                null, // Todo: Har vi en dagsats?
-                totalUtbetalingsgrad,
-                totalUtbetalingsgrad,
-                // Eksplisitt INFOTRYGD - v1 sin GrunnlagDto har kun Infotrygd som kilde. Viktig for
-                // riktig ruting til Infotrygd-mock fremfor Spøkelse-mock (default/null rutes til
-                // Spøkelse), se PersonTilGrunnlagMapper/PersonTilSykepengeVedtakMapper.
-                LegacyKilde.INFOTRYGD,
-                Collections.emptyList() // TODO: TREX til denne?
-        );
-    }
-
-    private static YtelseType tilYtelseTypeFraInfotrygdYtelse(GrunnlagDto.Ytelse ytelse) {
-        return switch (ytelse) {
-            case SP -> YtelseType.SYKEPENGER;
-            case BS -> YtelseType.PLEIEPENGER;
-        };
-    }
-
-    private static Ytelse tilYtelserFraPesys() {
-        return new Ytelse(YtelseType.UFØREPENSJON, LocalDate.now().minusYears(5), null, null, null, null, LegacyKilde.PESYS, null);
-    }
-
-    private static no.nav.vtp.person.personopplysninger.Språk tilSpråk(Språk språk) {
+    private static no.nav.vtp.person.personopplysninger.Språk tilSpråk(no.nav.foreldrepenger.vtp.kontrakter.person.v2.Språk språk) {
         if (språk == null) {
             return null;
         }
@@ -294,7 +126,7 @@ public class PersonMapper {
         };
     }
 
-    private static Kjønn tilKjønn(no.nav.foreldrepenger.vtp.kontrakter.person.Kjønn kjønn) {
+    private static Kjønn tilKjønn(no.nav.foreldrepenger.vtp.kontrakter.person.v2.Kjønn kjønn) {
         if (kjønn == null) {
             return null;
         }
@@ -308,10 +140,7 @@ public class PersonMapper {
         if (dto == null) {
             return null;
         }
-        return new GeografiskTilknytning(
-                dto.land(),
-                tilGeografiskTilknytningType(dto.type())
-        );
+        return new GeografiskTilknytning(Landkode.normaliser(dto.land()), tilGeografiskTilknytningType(dto.type()));
     }
 
     private static GeografiskTilknytning.GeografiskTilknytningType tilGeografiskTilknytningType(
@@ -319,38 +148,23 @@ public class PersonMapper {
         if (type == null) {
             return null;
         }
-        return switch (type) {
-            case BYDEL -> GeografiskTilknytning.GeografiskTilknytningType.BYDEL;
-            case KOMMUNE -> GeografiskTilknytning.GeografiskTilknytningType.KOMMUNE;
-            case LAND -> GeografiskTilknytning.GeografiskTilknytningType.LAND;
-        };
+        return GeografiskTilknytning.GeografiskTilknytningType.valueOf(type.name());
     }
 
-    private static List<Familierelasjon> tilFamilierelasjoner(List<FamilierelasjonModellDto> dtos, Map<UUID, PersonIdent> identer) {
+    private static List<Familierelasjon> tilFamilierelasjoner(List<FamilierelasjonDto> dtos, Map<UUID, PersonIdent> identer) {
         if (dtos == null) {
             return Collections.emptyList();
         }
         return dtos.stream()
-                .map(dto -> tilFamilierelasjon(dto, identer))
+                .map(dto -> new Familierelasjon(tilRelasjon(dto.relasjon()), identer.get(dto.relatertTilId())))
                 .toList();
     }
 
-    private static Familierelasjon tilFamilierelasjon(FamilierelasjonModellDto dto, Map<UUID, PersonIdent> identer) {
-        return new Familierelasjon(tilRelasjon(dto.relasjon()), identer.get(dto.relatertTilId()));
-    }
-
-    private static Familierelasjon.Relasjon tilRelasjon(FamilierelasjonModellDto.Relasjon relasjon) {
+    private static Familierelasjon.Relasjon tilRelasjon(FamilierelasjonDto.Relasjon relasjon) {
         if (relasjon == null) {
             return null;
         }
-        return switch (relasjon) {
-            case EKTE -> Familierelasjon.Relasjon.EKTE;
-            case SAMBOER -> Familierelasjon.Relasjon.SAMBOER;
-            case BARN -> Familierelasjon.Relasjon.BARN;
-            case FAR -> Familierelasjon.Relasjon.FAR;
-            case MOR -> Familierelasjon.Relasjon.MOR;
-            case MEDMOR -> Familierelasjon.Relasjon.MEDMOR;
-        };
+        return Familierelasjon.Relasjon.valueOf(relasjon.name());
     }
 
     private static List<Statsborgerskap> tilStatsborgerskap(List<StatsborgerskapDto> dtos) {
@@ -358,25 +172,21 @@ public class PersonMapper {
             return Collections.emptyList();
         }
         return dtos.stream()
-                .map(dto -> new Statsborgerskap(dto.land()))
-                .collect(Collectors.toList());
+                .map(dto -> new Statsborgerskap(Landkode.normaliser(dto.land())))
+                .toList();
     }
 
-    private static List<Sivilstand> tilSivilstand(List<SivilstandDto> dtos) {
+    private static List<Sivilstand> tilSivilstander(List<SivilstandDto> dtos) {
         if (dtos == null) {
             return Collections.emptyList();
         }
         return dtos.stream()
                 .map(PersonMapper::tilSivilstand)
-                .collect(Collectors.toList());
+                .toList();
     }
 
     private static Sivilstand tilSivilstand(SivilstandDto dto) {
-        return new Sivilstand(
-                tilSivilstander(dto.sivilstand()),
-                dto.fom(),
-                dto.tom()
-        );
+        return new Sivilstand(tilSivilstander(dto.sivilstand()), dto.fom(), dto.tom());
     }
 
     private static Sivilstand.Type tilSivilstander(SivilstandDto.Sivilstander sivilstand) {
@@ -387,9 +197,9 @@ public class PersonMapper {
             case ENKE -> Sivilstand.Type.ENKE_ELLER_ENKEMANN;
             case GIFT -> Sivilstand.Type.GIFT;
             case GJPA -> Sivilstand.Type.GJENLEVENDE_PARTNER;
-            case GLAD -> Sivilstand.Type.GIFT;  // GLAD (Giftet, lever adskilt) maps to GIFT
+            case GLAD -> Sivilstand.Type.GIFT;
             case REPA -> Sivilstand.Type.REGISTRERT_PARTNER;
-            case SAMB -> Sivilstand.Type.UGIFT;  // SAMB is not directly mapped, use UGIFT
+            case SAMB -> Sivilstand.Type.UGIFT;
             case SEPA -> Sivilstand.Type.SEPARERT_PARTNER;
             case SEPR -> Sivilstand.Type.SEPARERT;
             case SKIL -> Sivilstand.Type.SKILT;
@@ -403,35 +213,15 @@ public class PersonMapper {
             return Collections.emptyList();
         }
         return dtos.stream()
-                .map(PersonMapper::tilPersonstatus)
-                .collect(Collectors.toList());
-    }
-
-    private static Personstatus tilPersonstatus(PersonstatusDto dto) {
-        return new Personstatus(
-                tilPersonstatuser(dto.personstatus()),
-                dto.fom(),
-                dto.tom()
-        );
+                .map(dto -> new Personstatus(tilPersonstatuser(dto.personstatus()), dto.fom(), dto.tom()))
+                .toList();
     }
 
     private static Personstatus.Type tilPersonstatuser(PersonstatusDto.Personstatuser status) {
         if (status == null) {
             return null;
         }
-        return switch (status) {
-            case ABNR -> Personstatus.Type.ABNR;
-            case ADNR -> Personstatus.Type.ADNR;
-            case BOSA -> Personstatus.Type.BOSA;
-            case DØD -> Personstatus.Type.DØD;
-            case FOSV -> Personstatus.Type.FOSV;
-            case FØDR -> Personstatus.Type.FØDR;
-            case UFUL -> Personstatus.Type.UFUL;
-            case UREG -> Personstatus.Type.UREG;
-            case UTAN -> Personstatus.Type.UTAN;
-            case UTPE -> Personstatus.Type.UTPE;
-            case UTVA -> Personstatus.Type.UTVA;
-        };
+        return Personstatus.Type.valueOf(status.name());
     }
 
     private static List<Medlemskap> tilMedlemskap(List<MedlemskapDto> dtos) {
@@ -439,58 +229,34 @@ public class PersonMapper {
             return Collections.emptyList();
         }
         return dtos.stream()
-                .map(PersonMapper::tilMedlemskap)
-                .collect(Collectors.toList());
-    }
-
-    private static Medlemskap tilMedlemskap(MedlemskapDto dto) {
-        return new Medlemskap(
-                dto.fom(),
-                dto.tom(),
-                dto.land(),
-                tilDekningsType(dto.trygdedekning())
-        );
+                .map(dto -> new Medlemskap(dto.fom(), dto.tom(), Landkode.normaliser(dto.land()), tilDekningsType(dto.trygdedekning())))
+                .toList();
     }
 
     private static Medlemskap.DekningsType tilDekningsType(MedlemskapDto.DekningsType type) {
         if (type == null) {
             return null;
         }
-        return switch (type) {
-            case IHT_AVTALE -> Medlemskap.DekningsType.IHT_AVTALE;
-            case FULL -> Medlemskap.DekningsType.FULL;
-        };
+        return Medlemskap.DekningsType.valueOf(type.name());
     }
 
-    private static Adresser tilAdresser(List<AdresseDto> adresser, Adressebeskyttelse adressebeskyttelse) {
-        return new Adresser(tilAdresser(adresser), tilAdressebeskyttelse(adressebeskyttelse));
+    private static Adresser tilAdresser(List<AdresseDto> adresser, no.nav.foreldrepenger.vtp.kontrakter.person.v2.Adressebeskyttelse beskyttelse) {
+        var mapped = adresser == null ? List.<Adresse>of() : adresser.stream().map(PersonMapper::tilAdresse).toList();
+        return new Adresser(mapped, tilAdressebeskyttelse(beskyttelse));
     }
 
-    private static List<Adresse> tilAdresser(List<AdresseDto> adresser) {
-        return adresser.stream()
-                .map(PersonMapper::tilAdresse)
-                .toList();
-    }
-
-    private static Adresse tilAdresse(AdresseDto adresseDto) {
+    private static Adresse tilAdresse(AdresseDto dto) {
         return new Adresse(
-                tilAdresseType(adresseDto.adresseType()),
-                adresseDto.matrikkelId(),
-                adresseDto.land(),
-                adresseDto.fom(),
-                adresseDto.tom()
+                dto.adresseType() != null ? Adresse.AdresseType.valueOf(dto.adresseType().name()) : null,
+                dto.matrikkelId(),
+                Landkode.normaliser(dto.land()),
+                dto.fom(),
+                dto.tom()
         );
     }
 
-    private static Adresse.AdresseType tilAdresseType(AdresseDto.AdresseType adresseType) {
-        return switch (adresseType) {
-            case BOSTEDSADRESSE -> Adresse.AdresseType.BOSTEDSADRESSE;
-            case POSTADRESSE -> Adresse.AdresseType.POSTADRESSE;
-        };
-    }
-
     private static no.nav.vtp.person.personopplysninger.Adressebeskyttelse tilAdressebeskyttelse(
-            Adressebeskyttelse beskyttelse) {
+            no.nav.foreldrepenger.vtp.kontrakter.person.v2.Adressebeskyttelse beskyttelse) {
         if (beskyttelse == null) {
             return null;
         }
@@ -501,16 +267,43 @@ public class PersonMapper {
         };
     }
 
-    private static Arbeidsforholdstype tilArbeidsforholdstype(no.nav.foreldrepenger.vtp.kontrakter.person.Arbeidsforholdstype type) {
+    private static List<Arbeidsforhold> tilArbeidsforhold(List<ArbeidsforholdDto> arbeidsforhold, Map<UUID, PersonIdent> identer) {
+        if (arbeidsforhold == null) {
+            return Collections.emptyList();
+        }
+        return arbeidsforhold.stream()
+                .map(dto -> new Arbeidsforhold(
+                        tilArbeidsgiver(dto.arbeidsgiver(), dto.arbeidsforholdId(), identer),
+                        dto.ansettelsesperiodeFom(),
+                        dto.ansettelsesperiodeTom(),
+                        tilArbeidsforholdstype(dto.arbeidsforholdstype()),
+                        tilArbeidsavtaler(dto.arbeidsavtaler()),
+                        tilPermisjoner(dto.permisjoner())
+                ))
+                .toList();
+    }
+
+    private static no.nav.vtp.person.arbeidsforhold.Arbeidsgiver tilArbeidsgiver(ArbeidsgiverDto arbeidsgiver,
+                                                                                  String arbeidsforholdId,
+                                                                                  Map<UUID, PersonIdent> identer) {
+        if (arbeidsgiver instanceof OrganisasjonDto o) {
+            var detaljer = new Organisasjon.Detaljer(
+                    o.detaljer() != null ? o.detaljer().navn() : null,
+                    o.detaljer() != null ? o.detaljer().registreringsdato() : null
+            );
+            return new Organisasjon(new Orgnummer(o.orgnummer().value()), arbeidsforholdId, detaljer);
+        }
+        if (arbeidsgiver instanceof PrivatArbeidsgiverDto p) {
+            return new PrivatArbeidsgiver(identer.get(p.uuid()));
+        }
+        throw new IllegalStateException("Ukjent arbeidsgiver-type: " + arbeidsgiver);
+    }
+
+    private static Arbeidsforholdstype tilArbeidsforholdstype(no.nav.foreldrepenger.vtp.kontrakter.person.v2.Arbeidsforholdstype type) {
         if (type == null) {
             return null;
         }
-        return switch (type) {
-            case ORDINÆRT_ARBEIDSFORHOLD -> Arbeidsforholdstype.ORDINÆRT_ARBEIDSFORHOLD;
-            case MARITIMT_ARBEIDSFORHOLD -> Arbeidsforholdstype.MARITIMT_ARBEIDSFORHOLD;
-            case FRILANSER_OPPDRAGSTAKER_MED_MER -> Arbeidsforholdstype.FRILANSER_OPPDRAGSTAKER_MED_MER;
-            case FORENKLET_OPPGJØRSORDNING -> Arbeidsforholdstype.FORENKLET_OPPGJØRSORDNING;
-        };
+        return Arbeidsforholdstype.valueOf(type.name());
     }
 
     private static List<Arbeidsavtale> tilArbeidsavtaler(List<ArbeidsavtaleDto> dtos) {
@@ -518,19 +311,15 @@ public class PersonMapper {
             return Collections.emptyList();
         }
         return dtos.stream()
-                .map(PersonMapper::tilArbeidsavtale)
-                .collect(Collectors.toList());
-    }
-
-    private static Arbeidsavtale tilArbeidsavtale(ArbeidsavtaleDto dto) {
-        return new Arbeidsavtale(
-                dto.avtaltArbeidstimerPerUke(),
-                dto.stillingsprosent(),
-                dto.beregnetAntallTimerPerUke(),
-                dto.sisteLønnsendringsdato(),
-                dto.fomGyldighetsperiode(),
-                dto.tomGyldighetsperiode()
-        );
+                .map(dto -> new Arbeidsavtale(
+                        dto.avtaltArbeidstimerPerUke(),
+                        dto.stillingsprosent(),
+                        dto.beregnetAntallTimerPerUke(),
+                        dto.sisteLønnsendringsdato(),
+                        dto.fomGyldighetsperiode(),
+                        dto.tomGyldighetsperiode()
+                ))
+                .toList();
     }
 
     private static List<Permisjon> tilPermisjoner(List<PermisjonDto> dtos) {
@@ -538,53 +327,110 @@ public class PersonMapper {
             return Collections.emptyList();
         }
         return dtos.stream()
-                .map(PersonMapper::tilPermisjon)
-                .collect(Collectors.toList());
+                .map(dto -> new Permisjon(
+                        dto.fomGyldighetsperiode(),
+                        dto.tomGyldighetsperiode(),
+                        dto.stillingsprosent(),
+                        dto.permisjonstype() != null ? Permisjon.Permisjonstype.valueOf(dto.permisjonstype().name()) : null
+                ))
+                .toList();
     }
 
-    private static Permisjon tilPermisjon(PermisjonDto dto) {
-        return new Permisjon(
-                dto.fomGyldighetsperiode(),
-                dto.tomGyldighetsperiode(),
-                dto.stillingsprosent(),
-                tilPermisjonstype(dto.permisjonstype())
+    private static List<Ytelse> tilYtelser(List<YtelseDto> ytelser) {
+        if (ytelser == null) {
+            return Collections.emptyList();
+        }
+        return ytelser.stream()
+                .map(dto -> new Ytelse(
+                        YtelseType.valueOf(dto.type().name()),
+                        dto.fom(),
+                        dto.tom(),
+                        dto.dagsats(),
+                        // utbetalt (kroner) har ikke noe ekvivalent i v2-kontrakten - v2 ruter
+                        // alltid til nye mocker (Spøkelse, Kelvin, DPSAK).
+                        null,
+                        dto.utbetalingsgrad(),
+                        // V2-kontrakten har ikke kilde-felt. Alle v2-ytelser går til nye mocker;
+                        // vi lar kilde være null her og stoler på at nye mocker filtrerer bort
+                        // gamle kilder (LegacyKilde.ARENA/INFOTRYGD/PESYS) når de kalles.
+                        null,
+                        List.of()
+                ))
+                .toList();
+    }
+
+    private static List<Skatteopplysning> tilSkatteopplysninger(List<SkatteopplysningDto> skatteopplysninger) {
+        if (skatteopplysninger == null) {
+            return Collections.emptyList();
+        }
+        return skatteopplysninger.stream()
+                .map(dto -> new Skatteopplysning(dto.år(), dto.beløp()))
+                .toList();
+    }
+
+    /**
+     * Viderefører dagens mønster (Arena/Infotrygd i v1): alle ytelser genererer en Inntektsperiode,
+     * unntatt uføretrygd (samme unntak som Pesys-flyten har i dag). Beløpet kommer fra testenes
+     * egen dagsats — ingen hardkodede konstanter (i motsetning til v1s infotrygd()-hack med fast 10_000).
+     */
+    private static List<Inntektsperiode> tilInntekt(List<InntektsperiodeDto> inntekt,
+                                                    List<YtelseDto> ytelser,
+                                                    Map<UUID, PersonIdent> identer) {
+        var eksplisitt = inntekt == null ? Stream.<Inntektsperiode>empty()
+                : inntekt.stream().map(dto -> tilInntektsperiode(dto, identer));
+        var avledet = ytelser == null ? Stream.<Inntektsperiode>empty()
+                : ytelser.stream()
+                        .map(PersonMapper::tilInntektsperiodeFraYtelse)
+                        .filter(Objects::nonNull);
+        return Stream.concat(eksplisitt, avledet).toList();
+    }
+
+    private static Inntektsperiode tilInntektsperiode(InntektsperiodeDto dto, Map<UUID, PersonIdent> identer) {
+        return new Inntektsperiode(
+                tilArbeidsgiver(dto.arbeidsgiver(), null, identer),
+                dto.fom(),
+                dto.tom(),
+                dto.beløp(),
+                dto.ytelseType() != null ? Inntektsperiode.YtelseType.valueOf(dto.ytelseType().name()) : null,
+                dto.inntektFordel() != null ? Inntektsperiode.FordelType.valueOf(dto.inntektFordel().name()) : null
         );
     }
 
-    private static Permisjon.Permisjonstype tilPermisjonstype(Permisjonstype type) {
-        if (type == null) {
+    /** Uføretrygd unntas (som i dagens Pesys-flyt), og uten dagsats kan ingen inntektsperiode avledes. */
+    private static Inntektsperiode tilInntektsperiodeFraYtelse(YtelseDto dto) {
+        if (dto.type() == YtelseDto.YtelseType.UFØREPENSJON || dto.dagsats() == null) {
             return null;
         }
+        return new Inntektsperiode(
+                NAV_YTELSE_BETALING,
+                dto.fom(),
+                dto.tom(),
+                tilMånedligBeløp(dto.dagsats()),
+                tilInntektsperiodeYtelseType(dto.type()),
+                Inntektsperiode.FordelType.KONTANTYTELSE
+        );
+    }
+
+    /**
+     * Dagsats -> månedsbeløp basert på NAVs standard 260 virkedager per år. Inntektskomponent-mocken
+     * gjenbruker samme beløp for hver måned i perioden, så beløpet må representere en jevn månedlig
+     * inntekt uavhengig av periodens faktiske lengde.
+     */
+    private static Integer tilMånedligBeløp(Integer dagsats) {
+        return Math.round(dagsats * 260f / 12);
+    }
+
+    private static Inntektsperiode.YtelseType tilInntektsperiodeYtelseType(YtelseDto.YtelseType type) {
         return switch (type) {
-            case PERMISJON -> Permisjon.Permisjonstype.PERMISJON;
-            case PERMISJON_MED_FORELDREPENGER -> Permisjon.Permisjonstype.PERMISJON_MED_FORELDREPENGER;
-            case PERMISJON_VED_MILITÆRTJENESTE -> Permisjon.Permisjonstype.PERMISJON_VED_MILITÆRTJENESTE;
-            case PERMITTERING -> Permisjon.Permisjonstype.PERMITTERING;
-            case UTDANNINGSPERMISJON -> Permisjon.Permisjonstype.UTDANNINGSPERMISJON;
-            case UTDANNINGSPERMISJON_IKKE_LOVFESTET -> Permisjon.Permisjonstype.UTDANNINGSPERMISJON_IKKE_LOVFESTET;
-            case UTDANNINGSPERMISJON_LOVFESTET -> Permisjon.Permisjonstype.UTDANNINGSPERMISJON_LOVFESTET;
-            case VELFERDSPERMISJON -> Permisjon.Permisjonstype.VELFERDSPERMISJON;
-            case ANNEN_PERMISJON_IKKE_LOVFESTET -> Permisjon.Permisjonstype.ANNEN_PERMISJON_IKKE_LOVFESTET;
-            case ANNEN_PERMISJON_LOVFESTET -> Permisjon.Permisjonstype.ANNEN_PERMISJON_LOVFESTET;
-        };
-    }
-
-    private static Inntektsperiode.YtelseType tilYtelseType(InntektYtelseType type) {
-        if (type == null) {
-            return null;
-        }
-        // Map from DTO enum to domain enum - they should have matching values
-        return Inntektsperiode.YtelseType.valueOf(type.name());
-    }
-
-    private static Inntektsperiode.FordelType tilInntektFordel(InntektsperiodeDto.InntektFordelDto fordel) {
-        if (fordel == null) {
-            return null;
-        }
-        return switch (fordel) {
-            case KONTANTYTELSE -> Inntektsperiode.FordelType.KONTANTYTELSE;
-            case UTGIFTSGODTGJØRELSE -> Inntektsperiode.FordelType.UTGIFTSGODTGJØRELSE;
-            case NATURALYTELSE -> Inntektsperiode.FordelType.NATURALYTELSE;
+            case ARBEIDSAVKLARINGSPENGER -> Inntektsperiode.YtelseType.AAP;
+            case DAGPENGER -> Inntektsperiode.YtelseType.DAGPENGER;
+            case SYKEPENGER -> Inntektsperiode.YtelseType.SYKEPENGER;
+            case PLEIEPENGER -> Inntektsperiode.YtelseType.PLEIEPENGER;
+            case OMSORGSPENGER -> Inntektsperiode.YtelseType.OMSORGSPENGER;
+            case OPPLÆRINGSPENGER -> Inntektsperiode.YtelseType.OPPLÆRINGSPENGER;
+            case FORELDREPENGER -> Inntektsperiode.YtelseType.FORELDREPENGER;
+            case SVANGERSKAPSPENGER -> Inntektsperiode.YtelseType.SVANGERSKAPSPENGER;
+            case UFØREPENSJON -> throw new IllegalStateException("Uføretrygd skal være filtrert bort før dette punktet");
         };
     }
 }
