+++ server/src/test/java/no/nav/foreldrepenger/vtp/server/api/scenario/PersonMapperTest.java
@@ -3,24 +3,50 @@ package no.nav.foreldrepenger.vtp.server.api.scenario;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDate;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 
 import org.junit.jupiter.api.Test;
 
-import no.nav.foreldrepenger.vtp.kontrakter.person.Kjønn;
-import no.nav.foreldrepenger.vtp.kontrakter.person.Rolle;
-import no.nav.foreldrepenger.vtp.kontrakter.person.Språk;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.Kjønn;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.Rolle;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.Språk;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.AdresseDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.GeografiskTilknytningDto;
 import no.nav.foreldrepenger.vtp.kontrakter.person.v2.InntektsperiodeDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.MedlemskapDto;
 import no.nav.foreldrepenger.vtp.kontrakter.person.v2.OrganisasjonDto;
 import no.nav.foreldrepenger.vtp.kontrakter.person.v2.PersonDto;
 import no.nav.foreldrepenger.vtp.kontrakter.person.v2.PersonopplysningerDto;
+import no.nav.foreldrepenger.vtp.kontrakter.person.v2.StatsborgerskapDto;
 import no.nav.foreldrepenger.vtp.kontrakter.person.v2.YtelseDto;
 import no.nav.vtp.person.ident.PersonIdent;
 import no.nav.vtp.person.inntekt.Inntektsperiode;
+import no.nav.vtp.person.personopplysninger.Landkode;
 
-class PersonMapperV2Test {
+class PersonMapperTest {
+
+    @Test
+    void normalisererLandkoderTilAlpha3NårDomenemodellenOpprettes() {
+        var uuid = UUID.randomUUID();
+        var personopplysninger = PersonopplysningerDto.builder()
+                .uuid(uuid)
+                .geografiskTilknytning(new GeografiskTilknytningDto("no", GeografiskTilknytningDto.GeografiskTilknytningType.LAND))
+                .statsborgerskap(List.of(new StatsborgerskapDto("swe")))
+                .medlemskap(List.of(new MedlemskapDto(null, null, "DE", MedlemskapDto.DekningsType.FULL)))
+                .adresser(List.of(new AdresseDto(AdresseDto.AdresseType.BOSTEDSADRESSE, "nld", null, null, null)))
+                .build();
+        var dto = PersonDto.builder().personopplysninger(personopplysninger).build();
+
+        var person = PersonMapper.tilPerson(dto, Map.of(uuid, new PersonIdent("12345678901")), Optional.empty());
+
+        assertThat(person.personopplysninger().geografiskTilknytning().land()).isEqualTo(Landkode.NORGE);
+        assertThat(person.personopplysninger().statsborgerskap().getFirst().land()).isEqualTo("SWE");
+        assertThat(person.personopplysninger().medlemskap().getFirst().land()).isEqualTo("DEU");
+        assertThat(person.personopplysninger().adresser().adresser().getFirst().land()).isEqualTo("NLD");
+    }
 
     private static PersonDto.Builder personMedYtelse(UUID uuid, YtelseDto.YtelseType type, Integer dagsats) {
         return PersonDto.builder()
@@ -39,7 +65,7 @@ class PersonMapperV2Test {
         var uuid = UUID.randomUUID();
         var dto = personMedYtelse(uuid, YtelseDto.YtelseType.SYKEPENGER, 1_000).build();
 
-        var person = PersonMapperV2.tilPerson(dto, Map.of(uuid, new PersonIdent("12345678901")), Optional.empty());
+        var person = PersonMapper.tilPerson(dto, Map.of(uuid, new PersonIdent("12345678901")), Optional.empty());
 
         assertThat(person.inntekt()).hasSize(1);
         assertThat(person.inntekt().getFirst().ytelseType()).isEqualTo(Inntektsperiode.YtelseType.SYKEPENGER);
@@ -50,7 +76,7 @@ class PersonMapperV2Test {
         var uuid = UUID.randomUUID();
         var dto = personMedYtelse(uuid, YtelseDto.YtelseType.UFØREPENSJON, 1_000).build();
 
-        var person = PersonMapperV2.tilPerson(dto, Map.of(uuid, new PersonIdent("12345678901")), Optional.empty());
+        var person = PersonMapper.tilPerson(dto, Map.of(uuid, new PersonIdent("12345678901")), Optional.empty());
 
         assertThat(person.inntekt()).isEmpty();
         assertThat(person.ytelser()).hasSize(1);
@@ -61,7 +87,7 @@ class PersonMapperV2Test {
         var uuid = UUID.randomUUID();
         var dto = personMedYtelse(uuid, YtelseDto.YtelseType.DAGPENGER, null).build();
 
-        var person = PersonMapperV2.tilPerson(dto, Map.of(uuid, new PersonIdent("12345678901")), Optional.empty());
+        var person = PersonMapper.tilPerson(dto, Map.of(uuid, new PersonIdent("12345678901")), Optional.empty());
 
         assertThat(person.inntekt()).isEmpty();
     }
@@ -71,7 +97,7 @@ class PersonMapperV2Test {
         var uuid = UUID.randomUUID();
         var dto = personMedYtelse(uuid, YtelseDto.YtelseType.ARBEIDSAVKLARINGSPENGER, 1_000).build();
 
-        var person = PersonMapperV2.tilPerson(dto, Map.of(uuid, new PersonIdent("12345678901")), Optional.empty());
+        var person = PersonMapper.tilPerson(dto, Map.of(uuid, new PersonIdent("12345678901")), Optional.empty());
 
         var forventetMånedsbeløp = Math.round(1_000 * 260f / 12);
         assertThat(person.inntekt().getFirst().beløp()).isEqualTo(forventetMånedsbeløp);
@@ -82,7 +108,7 @@ class PersonMapperV2Test {
         var uuid = UUID.randomUUID();
         var dto = personMedYtelse(uuid, YtelseDto.YtelseType.PLEIEPENGER, 800).build();
 
-        var person = PersonMapperV2.tilPerson(dto, Map.of(uuid, new PersonIdent("12345678901")), Optional.empty());
+        var person = PersonMapper.tilPerson(dto, Map.of(uuid, new PersonIdent("12345678901")), Optional.empty());
 
         var arbeidsgiver = person.inntekt().getFirst().arbeidsgiver();
         assertThat(arbeidsgiver.identifikator()).isEqualTo("991013628");
@@ -96,7 +122,7 @@ class PersonMapperV2Test {
                 InntektsperiodeDto.YtelseType.FASTLØNN, InntektsperiodeDto.FordelType.KONTANTYTELSE);
         var dto = personMedYtelse(uuid, YtelseDto.YtelseType.DAGPENGER, null).inntekt(inntekt).build();
 
-        var person = PersonMapperV2.tilPerson(dto, Map.of(uuid, new PersonIdent("12345678901")), Optional.empty());
+        var person = PersonMapper.tilPerson(dto, Map.of(uuid, new PersonIdent("12345678901")), Optional.empty());
 
         assertThat(person.inntekt()).singleElement()
                 .extracting(Inntektsperiode::ytelseType)
